backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==4.1.0
boto3==1.28.9
botocore==1.31.9
celery==5.3.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-cors-headers==4.2.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-filter==23.2
django-js-asset==2.1.0
django-mptt==0.14.0
django-nine==0.2.7
django-phonenumber-field==7.1.0
django-redis==5.4.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.7
elastic-transport==8.4.1
elasticsearch==8.10.0
elasticsearch-dsl==8.9.0
factory-boy==3.3.0
Faker==19.6.2
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
install==1.3.5
jmespath==1.0.1
kombu==5.3.2
oauthlib==3.2.2
packaging==23.1
phonenumbers==8.13.11
phonenumberslite==8.13.11
Pillow==9.5.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
requests==2.30.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
transliterate==1.10.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.8
